Separated all the SignalR components
