+++ ClipboardMonitorLite.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipboardMonitorLite", "Cli
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SettingsLib", "SettingsLib\SettingsLib.csproj", "{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudMessageLib", "CloudMessageLib\CloudMessageLib.csproj", "{54171BE8-D12E-4406-9753-613AEF55385B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudConnectionLib", "CloudConnectionLib\CloudConnectionLib.csproj", "{971DF3CE-1EB6-41CF-9264-E02D38B62409}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Release|Any CPU.Build.0 = Release|Any CPU
-		{54171BE8-D12E-4406-9753-613AEF55385B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54171BE8-D12E-4406-9753-613AEF55385B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54171BE8-D12E-4406-9753-613AEF55385B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54171BE8-D12E-4406-9753-613AEF55385B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{971DF3CE-1EB6-41CF-9264-E02D38B62409}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{971DF3CE-1EB6-41CF-9264-E02D38B62409}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{971DF3CE-1EB6-41CF-9264-E02D38B62409}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{971DF3CE-1EB6-41CF-9264-E02D38B62409}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
