--- ClipboardMonitorLite/Cloud/OnlineState.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.SignalR.Client;
-
-namespace ClipboardMonitorLite.Cloud
-{
-    public static class OnlineState
-    {
-        public static HubConnectionState ConnectionLife = HubConnectionState.Disconnected;
-    }
-}
