+++ ClipboardMonitorLite/packages.config
@@ -1,29 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.AspNetCore.Connections.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.Http.Connections.Client" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.Http.Connections.Common" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.Http.Features" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.SignalR.Client" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.SignalR.Client.Core" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.SignalR.Common" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.SignalR.Protocols.Json" version="1.1.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Options" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Primitives" version="2.2.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
-  <package id="System.Buffers" version="4.5.0" targetFramework="net472" />
-  <package id="System.ComponentModel.Annotations" version="4.5.0" targetFramework="net472" />
-  <package id="System.IO.Pipelines" version="4.5.2" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.1" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.1" targetFramework="net472" />
-  <package id="System.Threading.Channels" version="4.5.0" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net472" />
 </packages>
