+++ CloudConnectionLib/OnlineState.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.SignalR.Client;
+
+namespace CloudConnectionLib
+{
+    public static class OnlineState
+    {
+        public static string ConnectionLife = "Disconnected";
+    }
+}
